feat: 랭크 조회 기능 구현

+++ AD/relpl/data/src/main/java/com/gdd/data/mapper/DataToDomain.kt
@@ -161,7 +161,7 @@ fun RankResponse.toRank():Rank {
 }
 
 fun RankResponseItem.toRankItem(): RankItem{
-    return RankItem(nickname, distance)
+    return RankItem(nickname, distance.toInt())
 }
